Refactor RemoveRun method in XWPFParagraph.cs
Updated the RemoveRun method to remove runs from the paragraph's low-level XML by searching the run to be removed in the CT_P Items list. The existing implementation didn't account for the existance of other element types in the low-level XML other than runs, thus the removal mechanism was unreliable. Besides the cycle to determine rPos that had no purpose, since rPos was not being used.